use nyaautils database api & remove mailbox

+++ src/main/java/cat/nyaa/HamsterEcoHelper/auction/AuctionInstance.java
@@ -73,7 +73,7 @@ public class AuctionInstance {
             } else {
                 realName += "(" + itemToGive.getType().name() + ":" + itemToGive.getDurability() + ")";
             }
-            int id = plugin.database.addItemLog(player, itemStack, startPrice, itemStack.getAmount());
+            long id = plugin.database.addItemLog(player, itemStack, startPrice, itemStack.getAmount());
             plugin.logger.info(I18n._("log.info.player_auc_start", id, player.getName(), realName, itemToGive.getAmount(),
                     Boolean.toString(hideName), startPrice, stepPrice, uid(this)));
         }
